catkin_ws/src/collision_avoidance/scripts/requirements.txt
gym==0.9.4
numpy==1.13.3
tensorflow-gpu==1.4.0
